Replace auto reload on wxt
