+++ entrypoints/background.js
@@ -1,0 +1,4 @@
+// noinspection ALL
+export default defineUnlistedScript(() => {
+    console.log("Background running...");
+});
