+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "sync-mate-extension",
   "version": "0.1.0",
-  "description": "",
+  "description": "Синхронизация видео между пользователями",
   "main": "index.js",
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack --mode production",
-    "dev": "cross-env NODE_ENV=development webpack --watch --mode development",
-    "lint": "eslint src tests --max-warnings=0",
-    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules npx jest"
+    "dev": "wxt",
+    "build": "wxt zip",
+    "lint": "eslint entrypoints tests --max-warnings=0",
+    "test": "cross-env NODE_OPTIONS=\"--experimental-vm-modules --no-warnings\" npx jest"
   },
   "repository": {
     "type": "git",
@@ -21,22 +21,14 @@
   },
   "homepage": "https://github.com/Zebaro24/Sync-Mate-Extension#readme",
   "devDependencies": {
-    "@babel/preset-env": "^7.28.3",
     "@eslint/js": "^9.33.0",
     "@eslint/json": "^0.13.1",
     "@jest/globals": "^30.0.5",
-    "babel-jest": "^30.0.5",
-    "copy-webpack-plugin": "^13.0.1",
+    "cross-env": "^10.0.0",
     "eslint": "^9.33.0",
     "globals": "^16.3.0",
     "jest": "^30.0.5",
     "jest-environment-jsdom": "^30.0.5",
-    "webpack": "^5.101.2",
-    "webpack-cli": "^6.0.1",
-    "webpack-ext-reloader": "^1.1.13"
-  },
-  "jest": {
-    "transform": {},
-    "testEnvironment": "jest-environment-jsdom"
+    "wxt": "^0.20.8"
   }
 }
