--- src/manifest.json
@@ -1,34 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Sync Mate",
-  "version": "0.1.0",
-  "description": "Синхронизация видео между пользователями",
-  "icons": {
-    "16": "assets/icon16.png",
-    "48": "assets/icon48.png",
-    "128": "assets/icon128.png"
-  },
-  "background": {
-    "service_worker": "background.js"
-  },
-  "content_scripts": [
-    {
-      "matches": [
-        "https://www.youtube.com/watch*",
-        "https://rezka.ag/*.html"
-      ],
-      "js": [
-        "content.js"
-      ],
-      "run_at": "document_end"
-    }
-  ],
-  "host_permissions": [
-    "https://www.youtube.com/watch*",
-    "https://rezka.ag/*.html"
-  ],
-  "permissions": [
-    "storage",
-    "scripting"
-  ]
-}
