+++ wxt.config.ts
@@ -1,0 +1,11 @@
+import {defineConfig} from "wxt";
+
+// noinspection JSUnusedGlobalSymbols
+export default defineConfig({
+    manifest: {
+        name: "Sync-Mate",
+    },
+    zip: {
+        artifactTemplate: 'Sync-Mate-Extension.zip',
+    },
+});
